Removing unnecessary-functions from `Account` implementations
New method for keeping track of the sorted accounts means we don't need to sort them on the fly.
We also don't query the accounts in this sorted way, either.

A `Config` is never displayed, so we can remove the `Display` trait.
It's a vestige from when I was troubleshooting a long time ago.