Convert tabs back to spaces.
Rust-analyzer uses hard-coded spaces as its indentation.
It's not because they prefer one way to another, but because all of the snippets are spaced out according to individual characters.
It's tough to do this same spacing with tabs in a generic way.

Since I use rust-analyzer as a coding assistant, and I don't require tabs everywhere, I'll just revert back to spaces.