Fixing issues with advancing ignored statements
Previously, advanced ignored statements when the dates were advanced, in `pair_dates_statements()`.
But the ignored statements are uncoupled from the actual dates, they only correspond to the statements being considered.
These should act like a stack, and be popped off the stack once they are used.
Moving the code to advance to the next iteration of the ignored statements to the function where it is used.