Fixing broken tests
By default, was creating a new statement with a blank path when pairing.
The default creation of a statement from `NaiveDate` didn't have this behvaiour, so there was a mismatch.
To solve this, I defined a `STATEMENT_DEFAULT_PATH_FMT` const that
`pair_dates_statements` and `Statement::from<&NaiveDate>` both use.

This ensures that both methods are creating empty statements in the same way so that tests will pass as expected.