Store the account statements within the CLI config
This will make it easier to aggregate all the TUI components under a single struct instead of passing `&acct_stmts` around everywhere.

To do this, needed to add `Clone` trait to `StatementCollection`, which is only used when refreshing the contents.
Also needed to duplicate `TryFrom` implementation for `StatementCollection` for `&mut Config` because it will be live-patched.

Finally, adding a few accessor and setter methods to `Config`.