Switch to `Config`'s internal statement collection to enable live reloading
A few rendering functions are altered to only require `Config` (and no longer require a separate `StatementCollection` as a parameter.

To enable live reloading with the `r` key, we needed to change some `&Config` parameters to `&mut Config`.
Then registering the `r` key with refreshing the account statements is the final piece to allow the user to reload the interface.

With manual testing this seems to be working perfectly well.