Implement `From<&IgnoredStatements>` for `IgnoreFile`
Now these two structs are reciprocal of each other.
The implementation here, to convert between them, is not my favourite based on the string formatting and parsing.
But there's no native interface here between the `toml` and `chrono` crates, so this is a quick and error-free way to do it.