Updating dependencies, removing oldtime feature on chrono
"oldtime" is a default feature provided by the chrono crate.
It depends on the `time` crate, which has a vulnerability associated with it.
`chrono` links to it for legacy purposes, but doesn't actively use it.

Because of this, I'm updating other dependencies and removing the "oldtime" feature (which is a default) from the dependencies.
This change affects all the subcrates here.